Fixes #149 

This PR implements in-memory caching of pgs response bodies inside Caddy using [a plugin](https://github.com/caddyserver/cache-handler) called [Souin](https://github.com/darkweak/souin).

### Background

Currently, serving a single asset from pgs can be a little slower than it could be due to DNS lookups, DB queries, and re-reading _headers and _redirects. Experiments in #151 showed that all these expensive operations are necessary and they cannot be simply cached in-process since pgs-web and pgs-ssh are separate processes that both need control over the cache. Thus, we have to use a cache like Souin which runs separate from pgs-web and pgs-ssh. But since it runs inside Caddy, no separate infrastructure is needed.

### How does the caching work?

When any user requests an asset like `https://example.com/styles.css` for the first time, the asset is fetched and returned to the user like normal, but now a copy of the response body and headers is stored by Souin (specifically using a high-performance in-memory store called [Otter](https://maypok86.github.io/otter/) which has impressive benchmarks and requires no special parameter tuning making it a good choice IMO). 

The cached response body is assigned a TTL of 5 minutes (which we could increase later if we get more confident with our cache flushing code). It's also associated with two keys:

* There's a main key which looks like `GET-https-example.com-/styles.css`. This key is used for quickly serving existing cached files when new requests arrive for the same assets.
* We also set a "[surrogate key](https://docs.fastly.com/en/guides/working-with-surrogate-keys)" which looks like `fakeuser-www-proj` (this matches the `*.pgs.sh` subdomain or the value in the TXT entry). This key is used for purging the cache for an entire project when any files are modified.

If any other users request the same file within 5 minutes, the responses will be served directly by Caddy from the in-memory cache. After 5 minutes, the cache for that asset expires.

Caveat: The cache is limited to 10,000 items. Otter uses [a special algorithm](https://s3fifo.com/) to evict rarely-accessed items from the cache to make room for new items.

### How does cache purging work?

Souin has an [API](https://github.com/darkweak/souin?tab=readme-ov-file#souin-api) for examining and purging entries from the cache. Unfortunately, there's [no way to expose this API on a different port](https://github.com/caddyserver/cache-handler/issues/106), so we have to protect it from abuse with Basic Auth.

When any files are written or deleted (like with [rsync](https://pico.sh/pgs#publish-your-site-with-one-command)), we use a golang channel to asynchronously purge the cache for the entire project by using [surrogate keys](https://docs.fastly.com/en/guides/working-with-surrogate-keys). This involves sending an HTTP request to `PURGE https://pgs.sh/souin-api/souin/`. These purge requests are debounced so that we only purge the cache once per site per 5 seconds.

This API is reachable from the public internet, just protected with basic auth. So in case of emergencies, admins can do things like `curl -i -X PURGE -u testuser:password https://pgs.sh/souin-api/souin/flush` to purge the entire cache.


### FAQ

* How much RAM will this use?
    - 50th percentile asset size seems to be around [500KB](https://almanac.httparchive.org/en/2021/page-weight) so a full cache (10,000 items) could be 5GB. Max theoretical is 10GB due to the 1MB `max_cacheable_body_bytes`.
* Are there any infrastructure changes required?
   - Mostly no. You'll need to rebuild/repush the Caddy image and you should generate a new basic_auth username/password.
* What isn't cached?
   - Errors, and the `/check` endpoint. Everything else (including the assets for https://pgs.sh) is cached.
* Are private pages still kept private?
    - I think so, since those requests bypass Caddy and go straight to `pgs-ssh`.

### Still TODO

* Actually execute this code. I tested individual parts but not the whole thing yet.
* Make sure this doesn't break other services like `imgs` or `prose`.
* **Figure out what to do about analytics.** Cached responses never hit pgs-web, so it can't increment counters in the DB. An easy/temp option could be not caching HTML files, but then we lose a lot of the benefits of caching. Possibly later we could pull page view counts from Caddy prometheus metrics.

Open to feedback and questions!