@mac-chaffee 2024-10-22
Was able to test this locally and found the surrogate key system doesn't behave [according to the spec](https://github.com/darkweak/souin/blob/master/pkg/surrogate/specification.md); they keys are never stored so the purging does nothing.

Everything else seems correct though.