@mac-chaffee 2024-10-22
Some benchmarking on my local laptop (I expect the performance difference to be even greater in production since my laptop has DNS caching and a very fast SSD):

### Without caching

```
$ oha -n 20000 -c 50 -t 10s --insecure -H "Host: picouser-www.localhost" https://localhost
Summary:
  Success rate:	100.00%
  Total:	64.1613 secs
  Slowest:	2.5051 secs
  Fastest:	0.0433 secs
  Average:	0.1604 secs
  Requests/sec:	311.7142

  Total data:	27.31 MiB
  Size/request:	1.40 KiB
  Size/sec:	435.86 KiB
```

### With caching
```
$ oha -n 20000 -c 50 -t 10s --insecure -H "Host: picouser-www.localhost" https://localhost
Summary:
  Success rate:	100.00%
  Total:	8.7945 secs
  Slowest:	2.4364 secs
  Fastest:	0.0005 secs
  Average:	0.0219 secs
  Requests/sec:	2274.1586
```