@neurosnap 2024-10-22
Thanks so much for taking the time to submit this PR!  I've been discussing it with my colleague @antoniomika 

One thing we are planning right now is multi-region pgs, I wonder how caching would work when we are serving sites for multiple regions (e.g. cache busting).

Further, we would also need to come up with a solution for analytics since that is a big feature for us.

Given these requirements, It makes us wonder if your original in-app cache would be better suited for our needs.

It kind of seems like we might want to implement pgs global first then think about a holistic caching system that can accommodate it.  Thoughts?