@mac-chaffee 2024-10-22
Multi-region pgs sounds like it would require the cache storage and the cache purging to also be multi-region too, right?

Do you all have a stance on the use of CDNs? IMO, "just swipe the credit card and use a CDN" would probably be easiest, but so far I've enjoyed the challenge of trying to find ways to avoid CDNs and additional infrastructure.

* Cloudflare supports caching and purging by tag with the header `Cache-Tag` (enterprise-only): https://developers.cloudflare.com/cache/how-to/purge-cache/purge-by-tags/
* Fastly supports caching and purging by tag with the header `Surrogate-Key`: https://docs.fastly.com/en/guides/working-with-surrogate-keys
* Akamai supports caching and purging by tag with the header `Edge-Cache-Tag`: https://techdocs.akamai.com/purge-cache/docs/assign-cache-tags
* You're running on Oracle Cloud, right? Doesn't look like their HTTP caching service supports tag-based cache purging: https://docs.oracle.com/en-us/iaas/Content/web-app-acceleration/acceleration-purge-cache.htm

With any of those options, we can mostly re-use the code in this PR, just removing the Caddy changes and changing the format of the API call in the `purgeCache` function (and the header name of course). Analytics would have to pulled from the CDN.

If you want to do multi-region caching without a CDN, one option is to simply replace Otter with one of the other [Souin storage backends](https://docs.souin.io/docs/storages/) that support clustering such as Redis. Still have to pull analytics from Caddy/Prometheus.

Another more-ambitious idea could be using/building a golang http caching middleware like https://github.com/victorspringer/http-cache , combined with something like https://github.com/riverqueue/river to coordinate cache purging through Postgres.

I don't have a strong opinion, I'm happy with any option that makes my pgs site blazing fast, whether it includes my Go code or not :)