@darkweak 2024-10-26
Hey, I'm the creator of Souin. First thank you for considering to add this HTTP cache in this great project.  
About the Cloudflare, Akamai, Fastly support, Souin can invalidate by tags and can store data in these CDNs.  
If you want a PG storage, I can try to find time to add this storage in the `storage` repository.

Open to discuss about your needs and what we could do together.