@neurosnap 2024-10-28
@mac-chaffee Thanks so much for taking the time to research and write a thoughtful response, much appreciated.

@darkweak Greetings!  Thanks for finding our little project and providing an awesome library (Souin), also much appreciated.

@antoniomika and I have had a chat and because of the way we record analytics and consider it a first-class feature at pico, we are leaning towards implementing the caching mechanism in our app code, although we can be convinced otherwise if there's a better way to outsource http caching (which is a solved problem that doesn't need re-inventing).

@mac-chaffee I like your idea of using `http-cache` since it simply wraps the Handler interface [Handler.ServeHTTP](https://cs.opensource.google/go/go/+/refs/tags/go1.23.2:src/net/http/server.go;l=88).  However, this doesn't solve the problem of being able to record analytics when there's a cache hit.  Further, we would ideally like to be standards compliant: [RFC 7234](https://datatracker.ietf.org/doc/html/rfc7234) which it appears `http-cache` isn't quite there yet.

Here's a patch that we could use to be able to record analytics for cache hits which we could try to upstream:
https://erock.pastes.sh/http-cache.diff

However, this still doesn't solve how we are going to clear the cache when a site gets uploaded since these are two separate services without a great way to communicate between them.  `river` is definitely a possibility.  Although, we could also use `http-cache.refreshKey` which would involve us sending an http request for each uri in a project.

Another important issue here is we support proxying requests to external services: https://pico.sh/pgs#proxy-to-another-service

We do *not* want to cache those responses.

In summary:
- We need to record site-usage analytics regardless of cache hits
- We must *not* cache http responses from proxy'd external services
- We need a way to clear cache keys when a user uploads new static files (which is in a separate service from the web service that uses the cache)

I'd be happy to continue facilitating this conversation so lemme know what you need from me!