@darkweak 2024-10-28
> However, this still doesn't solve how we are going to clear the cache when a site gets uploaded since these are two separate services without a great way to communicate between them.  river is definitely a possibility. Although, we could also use http-cache.refreshKey which would involve us sending an http request for each uri in a project.

You can tag your resources in Souin and invalidate by tag when needed, Souin exposes an API for that (you can purge by regexp, tag or per resource/uri).

> Another important issue here is we support proxying requests to external services: https://pico.sh/pgs#proxy-to-another-service. We do not want to cache those responses. If we went with http-cache we would need to figure out how to ignore caching some responses, ideally with a flag or something.

Just return a Cache-Control no-store directive should be enough for responses that are proxied.