@mac-chaffee 2024-10-28
Let's move the conversation to #149. I'll go ahead and close this but will leave the branch intact.