@neurosnap 2024-10-29
@darkweak Those are really nice features!  Do you have a recommendation on how we could still record cache hits in our site-usage analytics systems?